ft(api-endpoints): Use class for controllers
Implemented other feedback from LFA
