+++ Api/models/food.js
@@ -1,0 +1,14 @@
+const food = [{
+    foodId: 1,
+    foodName: 'CheeseBurger',
+    price: 1000
+  },
+
+  {
+    foodId: 2,
+    foodName: 'Fried Chicken',
+    price: 2000
+  }
+];
+
+export default food;
