+++ Api/models/user.js
@@ -1,0 +1,8 @@
+const user = [{
+  userId: 1,
+  name: 'Jacinta',
+  address: 'Awolowo Road Lagos',
+  phoneNumber: '08108765482'
+}];
+
+export default user;
