+++ README.md
@@ -1,10 +1,80 @@
 [![Build Status](https://travis-ci.org/Dubby20/Fast-Food-Fast.svg?branch=ch-integrate-travisCI-160483367)](https://travis-ci.org/Dubby20/Fast-Food-Fast)
 [![Coverage Status](https://coveralls.io/repos/github/Dubby20/Fast-Food-Fast/badge.svg?branch=ch-integrate-travisCI-160483367)](https://coveralls.io/github/Dubby20/Fast-Food-Fast?branch=ch-integrate-travisCI-160483367)
-[![Test Coverage](https://api.codeclimate.com/v1/badges/c5563c607a08e0628986/test_coverage)](https://codeclimate.com/github/Dubby20/Fast-Food-Fast/test_coverage)
 [![Maintainability](https://api.codeclimate.com/v1/badges/c5563c607a08e0628986/maintainability)](https://codeclimate.com/github/Dubby20/Fast-Food-Fast/maintainability)
+[![Test Coverage](https://api.codeclimate.com/v1/badges/c5563c607a08e0628986/test_coverage)](https://codeclimate.com/github/Dubby20/Fast-Food-Fast/test_coverage)
 
 
 
-
-# Fast-Food-Fast
+## Fast-Food-Fast
 Fast-Food-Fast is a food delivery service app for a restaurant
+
+
+## Features
++ Users can create account.
++ Users can log in.
++ Users can order for food.
++ Users can see a history of ordered food.
++ Admin can add food items.
++ Admin can edit food items.
++ Admin can delete food items.
++ Admin can see a list of food items.
++ Admin can see a list of orders.
++ Admin can accept or decline orders.
++ Admin can mark orders as completed.
+
+## Setup
+
+Step by step instructions on how to get the code setup locally. This may include:
++ Open the terminal
++ cd into directory that you want the project to reside.
+```
+cd projects
+```
++ clone the repository into that directory.
+```
+git clone https://github.com/Dubby20/Fast-Food-Fast.git
+```
+
+```
+run npm install && npm run start
+```
+
+### Dependencies
+
+List of libraries, tools, etc used for this project
+* [Nodejs](https://nodejs.org/en/)
+* [Express](https://expressjs.com/)
+<!-- * [Postgresql](https://www.postgresql.org/) -->
+* [Babel](https://babeljs.io) - Javascript compiler.
+* [Eslint](https://eslint.org/) - Javascript linter. [Airbnb](https://www.npmjs.com/package/eslint-config-airbnb) style [guide](https://github.com/airbnb/javascript)
+
+### Testing tools
+* [Mocha](https://mochajs.org/) - A Javascript test framework.
+* [Chai](http://chaijs.com) - A BDD / TDD Assertion library.
+* [Istanbul](https://istanbul.js.org) - Javascript code coverage tool.
+* [nyc](https://github.com/istanbuljs/nyc) - The Istanbul command line interface.
+
+## Documentation
+List of endpoints exposed by the service
+
+## Endpoints
+
+<table>
+<tr><th>Http verbs</th><th>Endpoints</th><th> Functionality </th></tr>
+<tr><td>GET</td><td>/orders </td><td> Gets all orderss</td></tr>
+<tr><td>GET</td><td>/orders/:orderId</td><td>Gets an order by id</td></tr>
+<tr><td>POST</td><td>/orders </td><td> Adds an order</td></tr>
+<tr><td>PUT</td><td> /orders/:orderId</td><td> Updates the status of an order</td></tr>
+</table>
+
+### Testing the application
+List of steps to run the service
+
+[Postman](www.getpostman.com)
+
+Running unit tests.
+* In a terminal, `cd` to the cloned project file.
+* Run `npm test`. This runs tests and displays coverage data generated by [Istanbul's](https://istanbul.js.org) nyc.
+
+## Author
+Jacinta Nnadi
