+++ package.json
@@ -14,12 +14,12 @@
   "scripts": {
     "mocha": "mocha --compilers js:babel-register",
     "babel-node": "babel-node --presets=env",
-    "start:dev": "nodemon --exec npm run babel-node -- api/server.js",
-    "start": "nodemon --exec npm run babel-node -- api/server.js",
+    "start:dev": "nodemon --exec npm run babel-node -- Api/server.js",
+    "start": "nodemon --exec npm run babel-node -- Api/server.js",
     "build": "babel ./ -d dist",
     "coveralls": "nyc npm test && nyc report --reporter=text-lcov | coveralls",
     "coverage": "nyc reports --reporter=text-lcov | coveralls",
-    "test": "npm run && nyc --reporter=text mocha  --timeout 100000 --compilers js:babel-register ./api/tests/*.js --exit --reporter spec"
+    "test": "npm run && nyc --reporter=text mocha  --timeout 100000 --compilers js:babel-register ./Api/tests/*.js --exit --reporter spec"
   },
   "repository": {
     "type": "git",
